Move i18n classes into subpackage and resource bundles into subfolder
