--- src/main/resources/strings_en.properties
@@ -1,6 +1,0 @@
-car-tableview-fxml-file=/fxml/CarTableView.fxml
-new-car-creationview-fxml-file=/fxml/NewCarCreationView.fxml
-car-configurationview-fxml-file=/fxml/CarConfigurationView.fxml
-employee-tableview-fxml-file=/fxml/EmployeeTableView.fxml
-rental-tableview-fxml-file=/fxml/RentalTableView.fxml
-customer-tableview-fxml-file=/fxml/CustomerTableView.fxml
