+++ src/test/java/de/htwsaar/prog3/carrental/i18n/I18nComponentsUtilTest.java
@@ -3,7 +3,7 @@ package de.htwsaar.prog3.carrental.i18n;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-import de.htwsaar.prog3.carrental.util.I18nComponentsUtil;
+import de.htwsaar.prog3.carrental.util.i18n.I18nComponentsUtil;
 
 import java.util.Locale;
